refactor utils to specific file, add some prelim jsdoc comments
