+++ src/applications/_mock-form-ae-design-patterns/vadx/server/routes/stop-on-port.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { killProcessOnPort } = require('../utils/utils');
+const { killProcessOnPort } = require('../utils/processes');
 const logger = require('../utils/logger');
 
 const router = express.Router();
